修正

+++ DirectXGame/Engine/Draw/2D/Sprite/Sprite2D.cpp
@@ -142,10 +142,10 @@ void Sprite2D::DrawAnimationSpriteXY2D(myMath::Vector2 position, uint16_t sheets
 	//頂点データ
 	PosUvColor vertices[] =
 	{
-		{{left,top,0.0f},{static_cast<float>(x / sheetsNumX),static_cast<float>(y / sheetsNumY)},{color.x, color.y, color.z, color.w}},//左上インデックス0
-		{{left,bottom,0.0f},{static_cast<float>(x / sheetsNumX),static_cast<float>((y + 1) / sheetsNumY)},{color.x, color.y, color.z, color.w}},//左下インデックス1
-		{{right,top,0.0f},{static_cast<float>((x + 1) / sheetsNumX),static_cast<float>(y / sheetsNumY)},{color.x, color.y, color.z, color.w}},//右上インデックス2
-		{{right,bottom,0.0f},{static_cast<float>((x + 1) / sheetsNumX),static_cast<float>((y + 1) / sheetsNumY)},{color.x, color.y, color.z, color.w}},//右下インデックス3
+		{{left,top,0.0f},{static_cast<float>(x) / static_cast<float>(sheetsNumX),static_cast<float>(y) / static_cast<float>(sheetsNumY)},{color.x, color.y, color.z, color.w}},//左上インデックス0
+		{{left,bottom,0.0f},{static_cast<float>(x) / static_cast<float>(sheetsNumX),(static_cast<float>(y) + 1) / static_cast<float>(sheetsNumY)},{color.x, color.y, color.z, color.w}},//左下インデックス1
+		{{right,top,0.0f},{(static_cast<float>(x) + 1) / static_cast<float>(sheetsNumX),static_cast<float>(y) / static_cast<float>(sheetsNumY)},{color.x, color.y, color.z, color.w}},//右上インデックス2
+		{{right,bottom,0.0f},{(static_cast<float>(x) + 1) / static_cast<float>(sheetsNumX),(static_cast<float>(y) + 1) / static_cast<float>(sheetsNumY)},{color.x, color.y, color.z, color.w}},//右下インデックス3
 	};
 
 	//インデックスデータ
